Logical Layer

+++ src/main/java/agh/kopec/evo_simulation/BehaviourManagerAlwaysNext.java
@@ -2,7 +2,7 @@ package agh.kopec.evo_simulation;
 
 public class BehaviourManagerAlwaysNext implements IBehaviourManager {
     @Override
-    public Integer chooseNewGene(Integer activeGene, Integer numberOfGenes) {
-        return (activeGene+1)%numberOfGenes;
+    public short chooseNewGene(short activeGene, short numberOfGenes) {
+        return (short) ((activeGene+1)%numberOfGenes);
     }
 }
